requirements.txt
pandas>=1.3.0
numpy>=1.21.0
matplotlib>=3.4.0
reportlab>=3.6.0
python-dotenv>=0.19.0
streamlit>=1.10.0
mssql-python
pyodbc>=4.0.0
sqlalchemy>=1.4.0
plotly>=5.3.0
kaleido>=0.2.1
pillow>=8.3.0
openpyxl>=3.0.0
tqdm>=4.62.0